Change namespace, prepare packaging and publishing
SDK and schema packages have been rename to `lumigator_*`.

Schemas are bundled with the SDK for external distribution, and packed
with the backend as source packages.

SDK packages are built using `uv` as part of CI and stored as
artifacts, which can be later pushed to Test PyPI in another CI
workflow.

Co-authored-by: Vicki Boykis <vicki@mozilla.ai>
Signed-off-by: javiermtorres <javier.maria.torres@gmail.com>