GHA: Rename the link checker GHA
Currently, we have two workflows named "Build and publish docs."
The first workflow builds the documentation using Sphinx and publishes
it to GitHub Pages, while the second checks for broken links. As a
result, there is no unique reference for either workflow.

Rename the link checker GitHub action to make it uniquely identifiable
for use in other workflows.

Refs #368

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>