[backend] Expose suggested models (#419)
* backend: Introduce models endpoint

Introduce a new `/models` endpoint that returns a list of suggested
models for a given task.

Currently, the only supported task is that of summarization.

Refs #381

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

* backend: Move suggested models in a YAML file

Move the suggested models list in a YAML file. Read the YAML file during
runtime to respond appropriately.

Refs #381

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

* models: Replace GPT-4-Turbo with GPT-4o

GPT-4-Turbo is a legacy OpenAI model.

Replace it with GPT-4o, which is the latest iteration of OpenAI's
flagship model.

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

* backend: Extend the models YAML file

Extend the models YAML file to include model-specific information.

Add model-specific details, such as the number of parameters, memory
footprint, and tensor type (e.g., FP32, BF16, etc.), along with a set
of default values used by Lumigator when calling the models.

Not all parameters are applicable to each model. For example, Hugging
Face models have default values for parameters like `max_length`, while
API models include parameters like `temperature`. In general, parameters
are model-specific, so there is no universal set of parameter names.

The defaults are inferred either by checking the documentation for each
model or by examining the configuration file on the Hugging Face Hub.

Refs #381

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

* backend: Get supported tasks from models YAML file

Use the models YAML file to retrieve the list of supported tasks,
instead of hardcoding it in the code.

Refs #381

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

* schemas: Validate Models Response object

Validate the response of the ``/models`` endpoint using Pydantic.

For each model, there are a few required fields:

* name
* uri
* description

There are also some optional fields, since not all models have available
information (e.g., we don't have the model size for OpenAI models) or
default parameters.

Finally, the default parameters dictionary is a generic `dict`, as there
is no guaranteed uniformity.

Refs #381

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

* backend: Test models endpoint

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

* sdk: Retrieve suggested models via the SDK

Retrieve the suggested models per task via the Lumigator SDK client.

Refs #381

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

* sdk: Test the models route

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

* docs: Add suggested models documentation

Add a page to the documentation listing the suggested models
and their details, along with instructions on how to retrieve
them using the API.

Closes #381

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

* docs: Fix SDK and Schema reference guide

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

---------

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>