backend: Watch inference job
Create a background task to monitor the progress of a submitted
inference job.

The FastAPI BackgroundTasks object is used to run a function in the
background. It is a wrapper arount Starlette's BackgroundTasks object. A
background task should be attached to a response, and will run only once
the response has been sent.

See:
* https://fastapi.tiangolo.com/tutorial/background-tasks/
* https://www.starlette.io/background/

Refs #405

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>