Fix results file name
When an inference or evaluation job completes successfully, it creates a
results file and stores it in object storage. The file is named either
`eval_results.json` or `inference_results.json`, depending on the job type.

However, when attempting to generate a download URL through
`/api/v1/jobs/{job_id}/result/download`, the URL always points to a file
named `eval_results.json`. For inference jobs, this file does not exist,
resulting in a broken download process.

Dynamically assign the file name based on the job type.

Refs #503

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>