backend: Standardize a job's output file name
When an inference or evaluation job completes successfully, it generates
a results file and stores it in object storage. The file is named either
`eval_results.json` or `inference_results.json`, based on the job type.

However, when generating a download URL via
`/api/v1/jobs/{job_id}/result/download`, the URL always points to a file
named `eval_results.json`. For inference jobs, this file does not exist,
leading to a broken download process.

Standardize the file name to `results.json`.

Refs #503

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>