Added /experiments_new endpoint
This commit adds a new route called `experiments_new`. The name is temporary and
it will be changed to `experiments` after we verify the functionality is the same.

The purpose of this new endpoint is to run an experiment by piping two simpler
jobs, one specific for inference and the other for evaluation.

At this stage, the code just replicates what is done by `experiments`.