Store the dataset generated by an inference job (#522)
* backend: Watch inference job

Create a background task to monitor the progress of a submitted
inference job.

The FastAPI BackgroundTasks object is used to run a function in the
background. It is a wrapper arount Starlette's BackgroundTasks object. A
background task should be attached to a response, and will run only once
the response has been sent.

See:
* https://fastapi.tiangolo.com/tutorial/background-tasks/
* https://www.starlette.io/background/

Refs #405

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

* backend: Upload inference-generated dataset

An inference job generates a `results.json` file that can be retrieved
and used to generate a new artificial dataset.

Leverage the running background task to retrieve `results.json`, extract
the necessary fields, create a CSV dataset in memory, and upload it to
the (local) database.

Refs #405

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

* backend: Add extra metadata to generated datasets

Datasets produced as a result of an inference job are considered
"generated". To ensure proper tracking and reliability:

- Store the run ID that generated the dataset for lineage purposes.
- Record the model used to generate the predictions to ensure
  traceability, even if Ray goes down.
- Annotate the dataset as "generated".

Refs #405

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

---------

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>