Allow current evaluation to skip inference
The main purpose of this commit is to have a quick way to test evaluation on
a dataset which already contains predictions, without depending on eval_lite
which is currently being implemented.

To do this, we add a `skip_inference` parameter to the *current* evaluation
(based on evaluator) which is optional and False by default. If this is True,
predictions are loaded from the input dataset (field named "predictions")
and summarization time is set to 0.0 (as we will lose this metric in the
eval_lite job and will gather it directly from the inference job results).

Updating the evaluation job configuration broke some tests which mistakenly
used the eval configuration for inference jobs, so they are fixed here too.

Also, we adapt the JobService so the dataset generated by inference has both
the ground_truth and the predictions fields, allowing us to run evaluation
directly on it.