#485 Build workflow calling inference + evaluation jobs sequentially (#540)
* Async jobs + _watch_job becomes _run_after_job accepting function

* Updated tests so only inference checks for mock.called

* Added /experiments_new entrypoint + implementation for mock 2-stages experiment

* Renamed prediction field to predictions in generated datasets

We usually refer to the field containing model-generated predictions as "predictions"
(see e.g. https://github.com/mozilla-ai/lumigator/blob/main/lumigator/python/mzai/jobs/evaluator/evaluator/jobs/evaluation/hf_evaluate.py#L146).

For consistency, we rename all occurrencies of "prediction" referring to this
field to "predictions".

* Allow current evaluation to skip inference

The main purpose of this commit is to have a quick way to test evaluation on
a dataset which already contains predictions, without depending on eval_lite
which is currently being implemented.

To do this, we add a `skip_inference` parameter to the *current* evaluation
(based on evaluator) which is optional and False by default. If this is True,
predictions are loaded from the input dataset (field named "predictions")
and summarization time is set to 0.0 (as we will lose this metric in the
eval_lite job and will gather it directly from the inference job results).

Updating the evaluation job configuration broke some tests which mistakenly
used the eval configuration for inference jobs, so they are fixed here too.

Also, we adapt the JobService so the dataset generated by inference has both
the ground_truth and the predictions fields, allowing us to run evaluation
directly on it.

* Run an actual inference+eval composite experiment

- the input is a dataset containing samples + ground truth
- the first job is inference, which generates a new dataset with an
  extra `predictions` field holding predictions by the chosen model
- the second job is evaluation, which takes the generated dataset
  as input and calculates the chosen metrics comparing predictions
  vs ground truth

The API interface for `/experiments_new` is the same as the old
`/experiment`, at least for the main POST, and the output is the
same format as the previous eval output.

The other endpoints (get experiment, list experiments, get results)
will be developed in the following commits.

* Flipped if

* Moved BackgroundTasks to experiments_new only

As ExperimentService currently has all the required information to run a task
such as `_add_dataset_to_db`, we decided to remove BackgroundTasks from the
JobService.

Main pros:
- JobService does not need to know anything about the logic required to run
  background tasks, same holds for the jobs route (and the current experiments
  route too, which matters less as it is going to be deprecated in favor of
  experiments_new)
- all the logic for piping different jobs/tasks is inside the ExperimentService
- testing individual jobs is not affected by BackgroundTasks (this will happen
  when we test experiments and we'll use mocking there)

Main cons:
- knowledge internal to individual jobs is not exposed to experiments, so we
  might need some workaround for that. One example is the default field name
  for model predictions in the inference job ("predictions"), which we have
  to explicitly set in the ExperimentService if we want the `add_dataset_to_db`
  method to know where to get predictions from

* Renamed `run_after_job` to `on_job_complete`

* Addressed reviews

---------

Signed-off-by: Davide Eynard <davide@mozilla.ai>