Integrate eval_lite as a new job
This commit integrates eval_lite as a separate, new job.
This is done to make sure that while we add this, everything still works
as expected. Once everything is tested and found in good status, we can
deprecate the older evaluator-based approach in favor of the two
inference + evaluation jobs and remove the extra code.

TODOs have been added with the purpose of finding the parts of code
we need to remove more easily